Factor out a common.ts.
